requirements.txt
anthropic>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-socketio>=5.3.0
flask-sqlalchemy>=3.1.0
pytest>=7.0.0
pytest-cov>=4.0.0